test app 98.97

+++ server/services/app/controllers/houseController.js
@@ -67,19 +67,12 @@ class HouseController {
   }
   static async updateHouse(req, res, next) {
     const t = await sequelize.transaction();
-    let { title, price, description, location, instalment, coordinate, Specifications, Images } =
+    let { title, price, description, location, instalment, coordinate, Specifications } =
       req.body;
     try {
-  
-      const house = await House.findByPk(req.params.id);
-      if (!house) {
-        throw {
-          name: "Not Found",
-          message: "House not found",
-        };
-      }
 
-      await House.update(
+    
+      const houseUpdate = await House.update(
         {
           title,
           price,
@@ -90,14 +83,33 @@ class HouseController {
         },
         { where: { id: req.params.id }, transaction: t }
       );
+      const findHouse = await House.findOne({
+        where:{
+          id: req.params.id
+        }
+      })
+
+      if (!findHouse) {
+        throw {
+          name: "Not Found",
+          message: "House not found",
+        };
+      }
+      let Images = req.uploadImages;
       Specifications.houseId = req.params.id;
-      Images.map((el) => (el.houseId = house.id));
+
+      Images.map((el) => (el.houseId = houseUpdate.id));
       await Specification.destroy({ where: { houseId: req.params.id }, transaction: t });
       await Image.destroy({ where: { houseId: req.params.id }, transaction: t });
       await Specification.create(Specifications, { transaction: t });
       await Image.bulkCreate(Images, { transaction: t });
       await t.commit();
-      res.status(200).json(house);
+     
+
+      res.status(200).json({
+        message: `House with id ${req.params.id} succesfully updated`,
+        data: findHouse
+      });
     } catch (err) {
       await t.rollback();
       next(err);
